+++ src/component.js
@@ -17,38 +17,21 @@ import { StreamWrapper } from './streamWrapper.js';
  * */
 export function tanokComponent(target) {
   target.propTypes = target.propTypes || {};
-  target.propTypes.eventStream = PropTypes.instanceOf(StreamWrapper);
   target.propTypes.tanokStream = PropTypes.instanceOf(StreamWrapper);
 
   target.displayName = `TanokComponent(${target.displayName || target.name})`;
 
   target.prototype.send = function send(action, payload, metadata = null) {
-    if (!this.props.tanokStream && this.props.eventStream) {
-      console.error(`Use 'tanokStream' argument instead of 'eventStream' (${target.displayName})`);
-    }
-
     if (metadata !== null) {
       console.error('Hey! You no longer can pass metadata `.send()`, use `.sub()`');
     }
 
-    const stream = this.props.tanokStream || this.props.eventStream;
+    const stream = this.props.tanokStream;
     stream.send(action, payload);
   };
 
-  target.prototype.subStream = function subStream(name, updateHandlers) {
-    console.error(`stream.subStream function is deprecated. Use subcomponentFx effect (${target.displayName})`);
-
-    const stream = this.props.tanokStream || this.props.eventStream;
-    return stream.subStream(name, updateHandlers);
-  };
-
   target.prototype.sub = function sub(name, metadata = null) {
-    if (!this.props.tanokStream && this.props.eventStream) {
-      console.error(`Use 'tanokStream' argument instead of 'eventStream' (${target.displayName})`);
-    }
-
-    const stream = this.props.tanokStream || this.props.eventStream;
-
+    const stream = this.props.tanokStream;
     return stream && stream.subWithMeta(`${stream.streamName}.${name}`, metadata);
   };